refactor: participate -> card/general 분리
