+++ src/components/surveys/participate/card/CardParticipate.jsx
@@ -1,14 +1,12 @@
 import React, { useState, useRef } from "react";
 import { useRecoilState } from "recoil";
-import { surveyForSubmit } from "../../../atoms";
+import { surveyForSubmit } from "../../../../atoms";
 import Dialog from "@mui/material/Dialog";
 import DialogActions from "@mui/material/DialogActions";
 import DialogContent from "@mui/material/DialogContent";
 import DialogContentText from "@mui/material/DialogContentText";
 import DialogTitle from "@mui/material/DialogTitle";
 import Button from "@mui/material/Button";
-import MultipleChoiceQuestionItem from "./MultipleChoiceQuestionItem";
-import PreviewSection from "../create/card/PreviewSection";
 import styled from "styled-components";
 import CardQuestionItem from "./CardQuestionItem";
