+++ src/components/surveys/participate/general/MultipleChoiceQuestionItem.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from "react";
-import { QuestionWrapper, QuestionOption, QuestionText } from "./styled";
+import { QuestionWrapper, QuestionOption, QuestionText } from "../styled";
 import {
   MultipleChoiceAnswerListState,
   multiChoiceAnswerValidateCount,
   surveyForSubmit,
-} from "../../../atoms";
+} from "../../../../atoms";
 import { useRecoilState, useRecoilValue } from "recoil";
 
 export default function MultipleChoiceQuestionItem({ item, multiquestion }) {
